clean-up of tools directory
