+++ .github/workflows/docgenerator.yml
@@ -65,7 +65,7 @@ jobs:
       - name: Release
         uses: softprops/action-gh-release@v1
         with:
-          # body_path: RECENT_CHANGES.md/RECENT_CHANGES.md
+          body_path: RECENT_CHANGES.md/RECENT_CHANGES.md
           # body_path: RECENT_CHANGES.md
           body: 'MSTG test'
           files: |
