--- tools/metadata.yml
@@ -1,9 +1,0 @@
----
-title: Mobile Security Testing Guide
-author: Bernhard Mueller, Sven Schleier, Jeroen Willemsen, Carlos Holguera, OWASP Mobile Team
-rights:  Creative Commons Attribution 4.0 International
-language: en-US
-tags: [Mobile Security, Android, iOS]
-abstract: |
-  This is the Mobile Security Testing Guide (MSTG).
----
