+++ tools/old/book.json
@@ -1,0 +1,13 @@
+{
+  "root" : "./Document",
+  "plugins" : [ "anchors" ],
+
+  "structure": {
+    "readme": "0x01-Foreword.md"
+  },
+  "title" : "OWASP Mobile Security Testing Guide []",
+  "language": "en",
+  "description" : "the MSTG is a comprehensive manual for mobile app security testing and reverse engineering for iOS and Android mobile security testers with the following content: 1. Mobile platform internals 2. Security testing in the mobile app development life cycle 3. Basic static and dynamic security testing 4. Mobile app reverse engineering and tampering 5. Assessing software protections 6. Detailed test cases that map to the requirements in the MASVS."
+
+
+}
